requirements.txt
Django==5.1.6
requests==2.32.3
python-dotenv==1.1.0